+++ src/main/java/DukeException.java
@@ -1,0 +1,6 @@
+package dav;
+class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
